+++ ostap/fitting/tests/test_fitting_dataset2.py
@@ -16,7 +16,9 @@ from   ostap.core.core          import dsID, hID, Ostap
 from   ostap.plotting.canvas    import use_canvas
 from   ostap.utils.progress_bar import progress_bar
 from   ostap.utils.memory       import memory 
-from   ostap.utils.root_utils   import batch_env 
+from   ostap.utils.root_utils   import batch_env
+from   ostap.logger.symbols     import delta_symbol, ram 
+from   ostap.fitting.dataset import data_ptr
 import ostap.logger.table       as     T
 import ostap.fitting.roofit 
 import ostap.trees.trees   
@@ -35,7 +37,6 @@ else :
 batch_env ( logger )
 # =============================================================================
 
-
 evt     = ROOT.RooRealVar ( 'Evt'    , '#event'        , 0 , 1000000 )
 run     = ROOT.RooRealVar ( 'Run'    , '#run'          , 0 , 1000000 )
 mass    = ROOT.RooRealVar ( 'Mass'   , 'mass-variable' , 0 , 100     )
@@ -46,12 +47,12 @@ weight  = ROOT.RooRealVar ( 'Weight' , 'some weight'   , -10 , 10 )
 varset  = ROOT.RooArgSet  ( evt , run , mass , pt1 , pt2 , weight )
 dataset = ROOT.RooDataSet ( dsID () , 'Test Data set-0' , varset )  
 
-NR  = 100
-NE  = 100
+NR  = 1000
+NE  =  200
 
 with memory ( 'Create initial dataset' , logger = logger ) as dm0 :
     
-    for r in progress_bar ( range ( NR ) )  :
+    for r in progress_bar ( range ( NR ) , description = "Create:")  :
         
         run.setVal ( r )
         for e in range ( NE ) :
@@ -75,38 +76,27 @@ logger.info ( 'Print initial dataset:\n%s' % dataset .table ( prefix = '# ' ) )
 
 
 
-from ostap.math.base import std
-data_ptr = std.unique_ptr(ROOT.RooAbsData)
-
 # =============================================================================
 ## (2)  Bootstrapping 
 # =============================================================================
 with memory ( 'Bootstrapping' , logger  = logger ) as dm1 :
 
-    size = 200
-    for i , ds in progress_bar ( enumerate ( dataset.bootstrap ( size  , extended = True ) ) ,
+    size = 10
+    for i , ds in progress_bar ( enumerate ( dataset.bootstrap ( size  , extended = True , wrap = True ) ) ,
                                  max_value   = size         , 
                                  description = "Bootstrap:" ) :
-        
-        ## ds = Ostap.MoreRooFit.delete_data ( ds )
-        ## ds.Delete()
-        ds = data_ptr ( ds )
-        del ds 
 
+        del ds 
 
 # =============================================================================
 ## (3) Jackknife  
 # =============================================================================
 with memory ( 'Jackknife' , logger  = logger ) as dm2 :
 
-    first , last = 0, 200 
-    for i , ds in progress_bar ( enumerate ( dataset.jackknife ( first , last ) ) ,
+    first , last = 0, 10
+    for i , ds in progress_bar ( enumerate ( dataset.jackknife ( first , last , wrap = True ) ) ,
                                  max_value   = 100          , 
                                  description = "Jackknife:" ) :
-
-        ## ds = Ostap.MoreRooFit.delete_data ( ds )
-        ## ds.Delete ()
-        ds = data_ptr ( ds ) 
         del ds 
 
 # ==============================================================================
@@ -115,9 +105,33 @@ with memory ( 'Delete Initial dataset' , logger = logger ) as dm3 :
     
     ## dataset = Ostap.MoreRooFit.delete_data ( dataset )
     ## del dataset 
-    dataset = data_ptr ( dataset )
+    ## dataset = data_ptr ( dataset )
     del dataset 
-    
+
+
+# ==============================================================================
+rows = [ ( "Action" , "%s %s memory [MB]" % ( delta_symbol , ram ) ) ] 
+
+row  = "Create"    , '%+.2f' % dm0.delta
+rows.append ( row ) 
+
+row  = "Bootstrap" , '%+.2f' % dm1.delta
+rows.append ( row ) 
+
+row  = "Jackknife" , '%+.2f' % dm2.delta
+rows.append ( row ) 
+
+row  = "Delete"    , '%+.2f' % dm3.delta
+rows.append ( row ) 
+
+title = "Memory usage"
+table = T.table ( rows , title = title , prefix = '# ' , alignment = 'lc' )
+logger.info ( '%s:\n%s' % ( title , table ) )
+
+assert dm1.delta < max ( 1 , 1.5 * dm0.delta ) , "There is some memory leak in bootstrap: %s vs %s " % ( dm1.delta , dm0.delta )
+assert dm2.delta < max ( 1 , 1.5 * dm0.delta ) , "There is some memory leak in jackknife: %s vs %s " % ( dm1.delta , dm0.delta )
+assert dm3.delta < max ( 1 , 1.5 * dm0.delta ) , "Memory is not released %s" % dm3.delta
+
 # =============================================================================
 ##                                                                      The END 
 # =============================================================================