+++ ostap/stats/ustat.py
@@ -240,7 +240,9 @@ def uToys ( pdf            ,
         counter += bool ( t_value < ti  )
 
         if isinstance ( ds , ROOT.RooDataSet ) :
-            ds = Ostap.MoreRooFit.delete_data ( ds )
+            ## ds = Ostap.MoreRooFit.delete_data ( ds )
+            from ostap.fitting.dataset import data_ptr 
+            ds = data_ptr ( ds )            
         del ds
         
     p_value = 1 - counter.eff